试卷的编辑

+++ plugins/examination/migrations/2020_12_21_164704_create_ims_yz_exam_paper_table.php
@@ -17,7 +17,7 @@ class CreateImsYzExamPaperTable extends Migration
                 $table->tinyInteger('random_answer', false, true)->default(0)->comment('选择乱序');
                 $table->string('random_topic', 255)->default('')->comment('随机选题:single,multiple,judgment,blank,qa');
                 $table->tinyInteger('question', false, true)->default(0)->comment('题目数量');
-                $table->tinyInteger('score', false, true)->default(0)->comment('总分值');
+                $table->smallInteger('score', false, true)->default(0)->comment('总分值');
                 $table->integer('created_at')->nullable();
                 $table->integer('updated_at')->nullable();
                 $table->integer('deleted_at')->nullable();
